Organize test code location
