+++ test/integration/presence_test.ts
@@ -1,6 +1,10 @@
 import { assert } from 'chai';
 import * as sinon from 'sinon';
-import yorkie, { DocEvent, DocEventType } from '@yorkie-js-sdk/src/yorkie';
+import yorkie, {
+  DocEvent,
+  DocEventType,
+  Counter,
+} from '@yorkie-js-sdk/src/yorkie';
 import {
   testRPCAddr,
   toDocKey,
@@ -537,5 +541,118 @@ describe(`Document.Subscribe('presence')`, function () {
     await c1.deactivate();
     await c2.deactivate();
     await c3.deactivate();
+  });
+});
+
+describe('Undo/Redo', function () {
+  it('Can undo/redo with presence', async function () {
+    type TestDoc = { counter: Counter };
+    type Presence = { cursor: { x: number; y: number } };
+    const docKey = toDocKey(`${this.test!.title}-${new Date().getTime()}`);
+    const doc = new yorkie.Document<TestDoc, Presence>(docKey);
+
+    const client = new yorkie.Client(testRPCAddr);
+    await client.activate();
+    await client.attach(doc, {
+      initialPresence: { cursor: { x: 0, y: 0 } },
+    });
+
+    doc.update((root, presence) => {
+      presence.set({ cursor: { x: 1, y: 1 } }, { addToHistory: true });
+    });
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 1, y: 1 },
+    });
+
+    doc.history.undo();
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 0, y: 0 },
+    });
+
+    doc.history.redo();
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 1, y: 1 },
+    });
+
+    doc.history.undo();
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 0, y: 0 },
+    });
+
+    doc.update((root) => {
+      root.counter = new Counter(yorkie.IntType, 100);
+    }, 'init counter');
+    assert.equal('{"counter":100}', doc.toSortedJSON());
+
+    doc.update((root, presence) => {
+      root.counter.increase(1);
+      presence.set({ cursor: { x: 2, y: 2 } }, { addToHistory: true });
+    }, 'increase 1');
+    assert.equal('{"counter":101}', doc.toSortedJSON());
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 2, y: 2 },
+    });
+
+    doc.history.undo();
+    assert.equal('{"counter":100}', doc.toSortedJSON());
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 0, y: 0 },
+    });
+
+    doc.history.redo();
+    assert.equal('{"counter":101}', doc.toSortedJSON());
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 2, y: 2 },
+    });
+
+    doc.history.undo();
+    assert.equal('{"counter":100}', doc.toSortedJSON());
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 0, y: 0 },
+    });
+
+    await client.deactivate();
+  });
+
+  it('Should not impact undo if presence is not added to history', async function () {
+    type TestDoc = { counter: Counter };
+    type Presence = { cursor: { x: number; y: number }; color: string };
+    const docKey = toDocKey(`${this.test!.title}-${new Date().getTime()}`);
+    const doc = new yorkie.Document<TestDoc, Presence>(docKey);
+
+    const client = new yorkie.Client(testRPCAddr);
+    await client.activate();
+    await client.attach(doc, {
+      initialPresence: { cursor: { x: 0, y: 0 }, color: 'red' },
+    });
+
+    doc.update((root, presence) => {
+      presence.set({ cursor: { x: 1, y: 1 } }, { addToHistory: true });
+      presence.set({ color: 'blue' });
+    });
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 1, y: 1 },
+      color: 'blue',
+    });
+
+    doc.history.undo();
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 0, y: 0 },
+      color: 'blue',
+    });
+
+    doc.history.redo();
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 1, y: 1 },
+      color: 'blue',
+    });
+
+    doc.history.undo();
+    assert.deepEqual(doc.getMyPresence(), {
+      cursor: { x: 0, y: 0 },
+      color: 'blue',
+    });
+
+    await client.deactivate();
   });
 });
